feat: add new storage
